uscheck_firestore/requirements.txt
qrcode
google-cloud-storage
markupsafe